str2bool moved
