+++ utils.py
@@ -7,6 +7,7 @@ import pandas as pd
 import random
 import numbers
 import torchvision
+import argparse
 
 from PIL import Image
 
@@ -295,3 +296,11 @@ def group_weight(weight_group, module, norm_layer, lr):
 	weight_group.append(dict(params=group_decay, lr=lr))
 	weight_group.append(dict(params=group_no_decay, weight_decay=.0, lr=lr))
 	return weight_group
+
+def str2bool(v):
+	if v.lower() in ('yes', 'true', 't', 'y', '1'):
+		return True
+	elif v.lower() in ('no', 'false', 'f', 'n', '0'):
+		return False
+	else:
+		raise argparse.ArgumentTypeError('Unsupported value encountered.')
